requirements.txt
keyboard==0.13.5
mss==9.0.1
pynput==1.7.6
pyperclip==1.8.2
google-cloud-vision==3.5.0
